move filter and datasource_dependency outside of worksheet_view_subelements as they can be found also elsewhere
